Rename dual single LLM shape to dual 1 to 1
