+++ scinoephile/lang/eng/ocr_fusion/prompts.py
@@ -8,7 +8,7 @@ from typing import ClassVar
 
 from scinoephile.core.text import dedent_and_compact
 from scinoephile.lang.eng.prompts import PromptEng
-from scinoephile.llms.dual_single.ocr_fusion import OcrFusionPrompt
+from scinoephile.llms.dual_1_to_1.ocr_fusion import OcrFusionPrompt
 
 __all__ = ["OcrFusionPromptEng"]
