+++ scinoephile/lang/zho/ocr_fusion/__init__.py
@@ -15,7 +15,7 @@ from scinoephile.llms.default_test_cases import (
     ZHO_HANT_OCR_FUSION_JSON_PATHS,
     load_default_test_cases,
 )
-from scinoephile.llms.dual_single.ocr_fusion import (
+from scinoephile.llms.dual_1_to_1.ocr_fusion import (
     OcrFusionManager,
     OcrFusionProcessor,
 )
